Address copy chat review feedback
