+++ src/components/content/HeaderGitBranchDropdown.vue
@@ -194,6 +194,7 @@ import IconTablerChevronDown from '../icons/IconTablerChevronDown.vue'
 import IconTablerFilePencil from '../icons/IconTablerFilePencil.vue'
 import IconTablerGitFork from '../icons/IconTablerGitFork.vue'
 import { useFeedbackDiagnostics } from '../../composables/useFeedbackDiagnostics'
+import { copyTextToClipboard } from '../../utils/clipboard'
 
 const props = defineProps<{
   currentBranch: string | null
@@ -340,28 +341,6 @@ function copyCommitRef(commit: GitCommitOption): void {
   void copyTextToClipboard(value).catch(() => {
     copiedCommitSha.value = ''
   })
-}
-
-async function copyTextToClipboard(value: string): Promise<void> {
-  try {
-    if (navigator.clipboard?.writeText) {
-      await navigator.clipboard.writeText(value)
-      return
-    }
-  } catch {
-    // Fall back for embedded browser contexts without clipboard permission.
-  }
-  const textarea = document.createElement('textarea')
-  textarea.value = value
-  textarea.setAttribute('readonly', 'true')
-  textarea.style.position = 'fixed'
-  textarea.style.top = '-9999px'
-  textarea.style.left = '-9999px'
-  document.body.appendChild(textarea)
-  textarea.select()
-  const copied = document.execCommand('copy')
-  textarea.remove()
-  if (!copied) throw new Error('Copy failed')
 }
 
 function resetSelectedCommit(): void {
